// src/types/common.enum.ts
export enum FilterUserType {
  ALL = "ALL",
  OPO = "OPO",
  PATHOLOGIST = "PATHOLOGIST",
  OTHER = "OTHER",
  TRANSPLANT = "TRANSPLANT_CENTER",
}
export enum UserRole {
  OPO = "OPO",
  PATHOLOGIST = "PATHOLOGIST",
  OTHER = "OTHER",
  TRANSPLANT = "TRANSPLANT_CENTER",
}
export enum DonorStatus {
  UNREGISTERED = "Unregistered",
  UNOS = "UNOS",
}

export enum OrganizationRole {
  LEADER = "LEADER",
  MEMBER = "MEMBER",
}